colors, borders, alignment
I changed all the background colors to the green that we had, also i got rid of all the borders, and finally i tried aligning all of the tables in the middle. I was able to do that, but the "Courses Remaining" text wont align center, i think because that is part of the PHP. Also the background color for the nav bar wont change for the same reason.